+++ src/main/kotlin/org/dnj/memoria/MemoriaExceptionHandler.kt
@@ -1,5 +1,6 @@
 package org.dnj.memoria
 
+import org.dnj.memoria.model.ErrorResponse
 import org.slf4j.LoggerFactory
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.ControllerAdvice
