+++ src/main/kotlin/org/dnj/memoria/database.kt
@@ -1,5 +1,8 @@
 package org.dnj.memoria
 
+import org.dnj.memoria.model.Item
+import org.dnj.memoria.model.Space
+import org.dnj.memoria.model.User
 import org.springframework.data.repository.CrudRepository
 
 interface UserRepository : CrudRepository<User, String> {
