+++ src/main/kotlin/org/dnj/memoria/model/login.kt
@@ -1,0 +1,11 @@
+package org.dnj.memoria.model
+
+data class LoginRequest(
+    val username: String,
+    val password: String
+)
+
+data class LoginResponse(
+    val username: String,
+    val jwt: String
+)
