+++ src/main/kotlin/org/dnj/memoria/service/AuthService.kt
@@ -3,9 +3,9 @@ package org.dnj.memoria.service
 import com.auth0.jwt.JWT
 import com.auth0.jwt.JWTVerifier
 import com.auth0.jwt.algorithms.Algorithm
-import org.dnj.memoria.LoginResponse
+import org.dnj.memoria.model.LoginResponse
 import org.dnj.memoria.MemoriaException
-import org.dnj.memoria.User
+import org.dnj.memoria.model.User
 import org.dnj.memoria.UserRepository
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.http.HttpStatus
